refactor: move specs to folder outer src
